added new component, and added useCallback for functions
